Ps.List.permutations
Currently returning entries for list sizes > 32 out of order
due to the particular [implementation of Maps within
Erlang](https://elixirforum.com/t/performance-of-tuples-over-maps/4591/5).

In the scheme of things, not a huge deal. Might also be helpful in
spotting better permutations more quickly. But, it's causing the
spec to fail, so it's "incomplete" at this point.