Add `Item.dig/2` and `~P` path sigil
`Item.dig/2` takes a complex data structure and a list of nodes,
consisting of strings, atoms, and integers, and traverses
through the data structure until either reaching the final given
node or reaching a node whose value is `nil`, and therefore no
longer traversable.

`~P` is syntactic sugar — in the form of a sigil — for various
functions within the `Path` module. A detailed explanation with
examples can be found in its documentation.

Also,
  - remove obsolete `Integer.is_odd` and `Integer.is_even` guards
  - use `import: true` for doctests to reduce boilerplate