Iterative, memory-bound, fast `exponential` algorithm
`Stream.unfold` is pretty cool. It's like a weird, inverse
corollary to `reduce`, and I just think that's nice.
Hoping that using `Enum.find` and folding in the work
toward the exponential's product will bring down the time
and memory usage… I mean, it should, since it means we're
not doing multiple different steps at each step in processing
the `exponent`, but then these benchmarks aren't the most
promising...